+++ app/languages/en_US.js
@@ -82,6 +82,8 @@ function en_USLang() {
     STR_EXIT_AGAIN_MULTI = "Click again to exit MultiStream!";
     STR_EXIT_MESSAGE = "Do you want to exit SmartTV Client for Twitch?";
     STR_EXIT = "Exit";
+    STR_CHANGELOG = "Changelog";
+    STR_CHANGELOG_SUMMARY = "Resume of the latest changes to read the full changelog check bellow link:";
     STR_CLOSE = "Close";
     STR_MINIMIZE = "Minimize";
     STR_CANCEL = "Cancel";