+++ app/specific/Sidepannel.js
@@ -333,6 +333,7 @@ function Sidepannel_KeyEnterBase() {
     else if (Sidepannel_Sidepannel_Pos === 12)
         Main_showControlsDialog(Sidepannel_Callback, ScreenObj[Screens_Current_Key].key_controls);
     else if (Sidepannel_Sidepannel_Pos === 13) Main_showExitDialog();
+    else if (Sidepannel_Sidepannel_Pos === 14) Main_showAboutDialog(Sidepannel_Callback, ScreenObj[Screens_Current_Key].key_controls, true);
 }
 
 function Sidepannel_KeyEnter() {
@@ -565,6 +566,7 @@ function Sidepannel_SetDefaultLables() {
     Main_innerHTML('side_panel_movel_new_11', STR_SPACE + STR_ABOUT);
     Main_innerHTML('side_panel_movel_new_12', STR_SPACE + STR_CONTROLS);
     Main_innerHTML('side_panel_movel_new_13', STR_SPACE + STR_EXIT);
+    Main_innerHTML('side_panel_movel_new_14', STR_SPACE + STR_CHANGELOG);
 
     Sidepannel_SetIcons('side_panel_new_1', 'search');
     Sidepannel_SetIcons('side_panel_new_2', 'user');
@@ -725,7 +727,7 @@ function Sidepannel_handleKeyDownMain(event) {
             break;
         case KEY_PG_DOWN:
         case KEY_DOWN:
-            if (Sidepannel_Sidepannel_Pos < 13) {
+            if (Sidepannel_Sidepannel_Pos < 14) {
                 Sidepannel_RemoveFocusMain();
                 Sidepannel_Sidepannel_Pos++;
                 if (!Main_values.Sidepannel_IsUser && Sidepannel_Sidepannel_Pos === 8) Sidepannel_Sidepannel_Pos += 2;